+++ utoipa/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+* Add support for `property_names` for object (https://github.com/juhaku/utoipa/pull/1084)
 * Add auto collect schemas for utoipa-axum (https://github.com/juhaku/utoipa/pull/1072)
 * Add global config for `utiopa` (https://github.com/juhaku/utoipa/pull/1048)
 * Add support for `links` in `#[utoipa::path]` (https://github.com/juhaku/utoipa/pull/1047)